requirements.txt
a2wsgi==1.7.0
amqp==5.1.1
anyio==3.7.0
aplus==0.11.0
asgiref==3.7.2
asteval==0.9.29
astropy==5.3
astroquery==0.4.6
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.1
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
blake3==0.3.3
blosc2==2.0.0
bqplot==0.12.39
branca==0.6.0
cachetools==5.3.1
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
-e git+ssh://git@github.com/django/channels.git@d5ff1f4bc159d2170a67e1ee237e0566fdf5bfa8#egg=channels
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
configparser==5.3.0
constantly==15.1.0
contourpy==1.0.7
cron-descriptor==1.4.0
cryptography==41.0.1
cycler==0.11.0
Cython==0.29.35
daphne==4.0.0
dask==2023.5.1
debugpy==1.6.7
decorator==5.1.1
Django==4.2.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-kombu==0.9.4
django-timezone-field==5.0
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.96.0
filelock==3.12.0
fonttools==4.39.4
frozendict==2.3.8
fsspec==2023.5.0
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
h11==0.14.0
h5py==3.8.0
habanero==1.2.3
hitran-api==1.2.2.0
hjson==3.1.0
html5lib==1.1
httptools==0.5.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
incremental==22.10.0
iniconfig==2.0.0
ipydatawidgets==4.3.3
ipykernel==6.23.1
ipyleaflet==0.17.3
ipympl==0.9.3
ipython==8.14.0
ipython-genutils==0.2.0
ipyvolume==0.6.3
ipyvue==1.9.1
ipyvuetify==1.8.10
ipywebrtc==0.6.0
ipywidgets==8.0.6
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json-tricks==3.17.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
keyring==23.13.1
kiwisolver==1.4.4
kombu==5.3.0
llvmlite==0.40.0
lmfit==1.2.1
locket==1.0.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==9.1.0
mpldatacursor==0.7.1
msgpack==1.0.5
mysqlclient==2.1.1
nest-asyncio==1.5.6
numba==0.57.0
numexpr==2.8.4
numpy==1.22.3
packaging==23.1
pandas==2.0.2
parso==0.8.3
partd==1.4.0
PeakUtils==1.3.4
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.15.0
pluggy==1.0.0
progressbar2==4.2.0
prompt-toolkit==3.0.38
psutil==5.9.5
publib==0.3.3
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
pyerfa==2.0.0.3
Pygments==2.15.1
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pytest==7.3.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-utils==3.6.0
pythreejs==2.4.2
pytz==2023.3
pyvo==1.4.1
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==25.1.0
radis==0.14
redis==4.5.5
requests==2.31.0
rich==13.4.1
scipy==1.10.1
seaborn==0.12.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
tables==3.8.0
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
traittypes==0.2.1
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
uncertainties==3.1.7
urllib3==2.0.3
uvicorn==0.22.0
vaex==4.16.0
vaex-astro==0.9.3
vaex-core==4.16.1
vaex-hdf5==0.14.1
vaex-jupyter==0.8.1
vaex-ml==0.18.1
vaex-server==0.8.1
vaex-viz==0.5.4
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
widgetsnbextension==4.0.7
xarray==2023.5.0
xyzservices==2023.5.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
